Gather error class into one place
And ensure these errors inherit same base error class.
